Changed data types used in these tests/examples to be those included in Tpetra's ETI.

+++ packages/zoltan2/example/block/block.cpp
@@ -71,7 +71,7 @@ int main(int argc, char *argv[])
 #endif
 
   typedef double scalar_t;
-  typedef long   localId_t;
+  typedef int    localId_t;
   typedef long   globalId_t;
 
   ///////////////////////////////////////////////////////////////////////
