+++ packages/zoltan2/example/geometric/rcb_C.cpp
@@ -72,9 +72,8 @@ int main(int argc, char *argv[])
   int rank=0, nprocs=1;
 #endif
 
-  // TODO explain
-  typedef float  scalar_t;
-  typedef int    localId_t;
+  typedef double scalar_t;
+  typedef int localId_t;
   typedef long globalId_t;
 
   typedef Zoltan2::BasicUserTypes<scalar_t, globalId_t, localId_t, globalId_t> myTypes;
